pyproject.toml
[project]
name = "whisper-diarization-demo"
version = "0.1.0"
description = "An offline-first demo for Chinese speaker diarization and ASR, powered by pyannote.audio and OpenAI Whisper"
readme = "README.md"
requires-python = ">=3.9.12"
license = { text = "MIT" }
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
keywords = [
    "speech-recognition",
    "speaker-diarization", 
    "whisper",
    "pyannote",
    "chinese",
    "offline",
    "asr"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "openai-whisper>=20231117",
    "pyannote.audio>=3.1.0",
    "pydub>=0.25.1",
    "torch>=2.1.0,<2.2.0",
    "torchaudio>=2.1.0,<2.2.0",
    "python-dotenv>=1.0.0",
    "huggingface_hub<1.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pre-commit>=3.5.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/whisper-diarization-demo"
Repository = "https://github.com/yourusername/whisper-diarization-demo"
Issues = "https://github.com/yourusername/whisper-diarization-demo/issues"

[project.scripts]
whisper-diarization = "whisper_diarization.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/whisper_diarization"]

# Ruff 配置 - 代码检查和格式化
[tool.ruff]
line-length = 100
target-version = "py39"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # Pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py
"__main__.py" = ["E402"]  # module level import not at top (需要先设置 warnings)
"config.py" = ["E402"]    # module level import not at top (需要先加载环境变量)

# Pytest 配置
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--cov=src/whisper_diarization",
    "--cov-report=term-missing",
    "--cov-report=html",
]

# Mypy 配置 - 类型检查
[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true

[[tool.mypy.overrides]]
module = [
    "pyannote.*",
    "whisper.*",
    "pydub.*",
]
ignore_missing_imports = true

# Coverage 配置
[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

